Fixed or suppressed as appropriate a double handful of warnings new in XCode 5.1.

+++ Frameworks/AIUtilities Framework/Source/AIDateFormatterAdditions.m
@@ -353,11 +353,11 @@ static AIDateFormatterCache *sharedFormatterCache = nil;
 		minutesString	= [NSString stringWithFormat: @"%lim",minutes];
 		secondsString	= [NSString stringWithFormat: @"%.0fs",seconds];
 	} else {
-		weeksString		= (weeks == 1)		? ONE_WEEK		: [NSString stringWithFormat:MULTIPLE_WEEKS, weeks];
-		daysString		= (days == 1)		? ONE_DAY		: [NSString stringWithFormat:MULTIPLE_DAYS, days];
-		hoursString		= (hours == 1)		? ONE_HOUR		: [NSString stringWithFormat:MULTIPLE_HOURS, hours];
-		minutesString	= (minutes == 1)	? ONE_MINUTE	: [NSString stringWithFormat:MULTIPLE_MINUTES, minutes];
-		secondsString	= (seconds == 1)	? ONE_SECOND	: [NSString stringWithFormat:MULTIPLE_SECONDS, seconds];
+		weeksString		= (weeks == 1)		? ONE_WEEK		: [NSString stringWithFormat:MULTIPLE_WEEKS, (int)weeks];
+		daysString		= (days == 1)		? ONE_DAY		: [NSString stringWithFormat:MULTIPLE_DAYS, (int)days];
+		hoursString		= (hours == 1)		? ONE_HOUR		: [NSString stringWithFormat:MULTIPLE_HOURS, (int)hours];
+		minutesString	= (minutes == 1)	? ONE_MINUTE	: [NSString stringWithFormat:MULTIPLE_MINUTES, (int)minutes];
+		secondsString	= (seconds == 1)	? ONE_SECOND	: [NSString stringWithFormat:MULTIPLE_SECONDS, (int)seconds];
 	}
 
 	//assemble the parts
